Decided to move the Layers and _Animation classes into seperate files, to avoid confusion with the Timeline class.
