+++ src/Layers.js
@@ -1,35 +1,6 @@
 import Drawing from "./Drawing.js";
+import _Animation from "./_Animation.js";
 import { mod } from "./utils.js";
-
-class _Animation {
-    /**@type {number}*/ static length = null;
-    static #frame = 0;
-    static set frame(frame) {
-        if (this.length === null) {
-            throw new Error("_Animation.length must be set before modifying _Animation.frame.");
-        }
-        this.#frame = mod(frame, this.length);
-    }
-    static get frame() { return this.#frame }
-
-    /** @type {Drawing[]} */
-    #animation;
-
-    /**  @param {string} fillCharacter @param {Drawing[]} animation*/
-    constructor(fillCharacter, animation = null) {
-        if (_Animation.length === null) {
-            throw new Error("_Animation.length must be set before creating an _Animation object.");
-        }
-        this.#animation = animation !== null ? animation :
-            this.#animation = new Array(_Animation.length).fill()
-                .map(x => new Drawing(fillCharacter));
-    }
-
-    get current() { return this.#animation[_Animation.frame] }
-    set current(drawing) { this.#animation[_Animation.frame] = drawing }
-
-    get animation() { return this.#animation }
-}
 
 class Layers {
     initTimeline() {
@@ -59,7 +30,7 @@ class Layers {
 
     /**  @param {string} fillCharacter @param {(_Animation | Drawing)[]} layers*/
     constructor(fillCharacter, layers = null) {
-        if (_Animation.Glength === null) {
+        if (_Animation.length === null) {
             throw new Error("Layers.length must be set before creating a Layers object.");
         }
         if (layers !== null) {
@@ -76,4 +47,4 @@ class Layers {
     set current(drawing) { this.#layers[Layers.#layer].current = drawing }
 }
 
-export { _Animation, Layers }
+export default Layers;
